Work on NIF support: able to load marker_arrow.nif
Fixed existing NIF block definitions, and added missing ones.

`nifDK::file_reader::read_indexed_string` now supports string index -1, which is apparently used in some files to indicate the absence of a string. Additionally, out-of-bounds string indices now throw.

We are now able to accurately load `marker_arrow.nif`.

+++ native/src/DovahKit/nif/blocks/NiGeometryData.h
@@ -35,7 +35,7 @@ namespace nifDK::block_types {
          uint32_t material_crc = 0;
          NiBound bounds;
          std::array<uint16_t, 13> unknown_shorts = {};
-         uint32_t consistency_flags = 0;
+         uint16_t consistency_flags = 0;
          NiObject* additional = nullptr;
 
          virtual void parse(file_reader&) override;
